+++ src/test/groovy/fr/nereide/test/inspection/XmlDuplicateInspectionTest.groovy
@@ -6,19 +6,16 @@ import fr.nereide.inspection.xml.DuplicatedFormInspection
 import fr.nereide.inspection.xml.DuplicatedMenuInspection
 import fr.nereide.inspection.xml.DuplicatedScreenInspection
 import fr.nereide.inspection.xml.DuplicatedServiceXmlInspection
-import fr.nereide.inspection.xml.DuplicatedUriInspection
 
 import static fr.nereide.inspection.InspectionBundle.message
 
 class XmlDuplicateInspectionTest extends BaseInspectionTest {
 
-    final String URI_DESC = message('inspection.uri.duplicate.display.descriptor')
     final String FORM_DESC = message('inspection.form.duplicate.display.descriptor')
     final String SCREEN_DESC = message('inspection.screen.duplicate.display.descriptor')
     final String MENU_DESC = message('inspection.menu.duplicate.display.descriptor')
     final String SERVICE_DESC = message('inspection.service.duplicate.display.descriptor')
     final String ENTITY_DESC = message('inspection.entity.duplicate.display.descriptor')
-    final LocalInspectionTool URI_INSP = new DuplicatedUriInspection()
     final LocalInspectionTool FORM_INSP = new DuplicatedFormInspection()
     final LocalInspectionTool SCREEN_INSP = new DuplicatedScreenInspection()
     final LocalInspectionTool MENU_INSP = new DuplicatedMenuInspection()
@@ -36,8 +33,6 @@ class XmlDuplicateInspectionTest extends BaseInspectionTest {
     protected void setUp() {
         super.setUp()
         String myTestFileName = "${this.getTestName(false)}.xml"
-        myFixture.createFile(myTestFileName, '')
-        myFixture.moveFile(myTestFileName, DEST)
         myFixture.copyFileToProject("xml/$myTestFileName", "$DEST/$myTestFileName")
     }
 
@@ -46,12 +41,6 @@ class XmlDuplicateInspectionTest extends BaseInspectionTest {
         String file = "${this.getTestName(false)}.xml"
         myFixture.configureByFile("$DEST/$file")
         doHighlightTest(mustFind, desc)
-    }
-
-    void testDuplicatedUriInCurrentController() { doTest(URI_DESC, URI_INSP) }
-
-    void testDuplicatedUriInCurrentControllerWithDifferentMethod() {
-        doTest(false, URI_DESC, URI_INSP)
     }
 
     void testDuplicatedScreenInCurrentFile() { doTest(SCREEN_DESC, SCREEN_INSP) }
@@ -83,4 +72,5 @@ class XmlDuplicateInspectionTest extends BaseInspectionTest {
     void testDuplicatedEntityIsCalled() { doTest(ENTITY_DESC, ENTITY_INSP) }
 
     void testDuplicatedViewIsCalled() { doTest(ENTITY_DESC, ENTITY_INSP) }
+
 }